requirements-runtime.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra=runtime --output-file=requirements-runtime.txt --strip-extras --unsafe-package=distribute --unsafe-package=localstack-core --unsafe-package=pip --unsafe-package=setuptools pyproject.toml
#
airspeed-ext==0.6.7
    # via localstack-core (pyproject.toml)
amazon-kclpy==3.0.1
    # via localstack-core (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
antlr4-python3-runtime==4.13.2
    # via
    #   localstack-core (pyproject.toml)
    #   moto-ext
apispec==6.8.0
    # via localstack-core (pyproject.toml)
argparse==1.4.0
    # via amazon-kclpy
attrs==24.3.0
    # via
    #   jsonschema
    #   localstack-twisted
    #   referencing
aws-sam-translator==1.94.0
    # via
    #   cfn-lint
    #   localstack-core (pyproject.toml)
aws-xray-sdk==2.14.0
    # via moto-ext
awscli==1.36.27
    # via localstack-core (pyproject.toml)
awscrt==0.23.6
    # via localstack-core
boto3==1.35.86
    # via
    #   amazon-kclpy
    #   aws-sam-translator
    #   localstack-core
    #   moto-ext
botocore==1.35.86
    # via
    #   aws-xray-sdk
    #   awscli
    #   boto3
    #   localstack-core
    #   moto-ext
    #   s3transfer
build==1.2.2.post1
    # via
    #   localstack-core
    #   localstack-core (pyproject.toml)
cachetools==5.5.0
    # via
    #   airspeed-ext
    #   localstack-core
    #   localstack-core (pyproject.toml)
cbor2==5.6.5
    # via localstack-core
certifi==2024.12.14
    # via
    #   opensearch-py
    #   requests
cffi==1.17.1
    # via cryptography
cfn-lint==1.22.2
    # via moto-ext
charset-normalizer==3.4.0
    # via requests
click==8.1.8
    # via
    #   localstack-core
    #   localstack-core (pyproject.toml)
colorama==0.4.6
    # via awscli
constantly==23.10.4
    # via localstack-twisted
crontab==1.0.1
    # via localstack-core (pyproject.toml)
cryptography==44.0.0
    # via
    #   joserfc
    #   localstack-core
    #   localstack-core (pyproject.toml)
    #   moto-ext
    #   pyopenssl
decorator==5.1.1
    # via jsonpath-rw
dill==0.3.6
    # via
    #   localstack-core
    #   localstack-core (pyproject.toml)
dnslib==0.9.25
    # via
    #   localstack-core
    #   localstack-core (pyproject.toml)
dnspython==2.7.0
    # via
    #   localstack-core
    #   localstack-core (pyproject.toml)
    #   pymongo
docker==7.1.0
    # via
    #   localstack-core
    #   moto-ext
docutils==0.16
    # via awscli
events==0.5
    # via opensearch-py
graphql-core==3.2.5
    # via moto-ext
h11==0.14.0
    # via
    #   hypercorn
    #   wsproto
h2==4.1.0
    # via
    #   hypercorn
    #   localstack-twisted
hpack==4.0.0
    # via h2
hypercorn==0.17.3
    # via localstack-core
hyperframe==6.0.1
    # via h2
hyperlink==21.0.0
    # via localstack-twisted
idna==3.10
    # via
    #   hyperlink
    #   localstack-twisted
    #   requests
incremental==24.7.2
    # via localstack-twisted
isodate==0.7.2
    # via openapi-core
jinja2==3.1.5
    # via moto-ext
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joserfc==1.0.1
    # via moto-ext
jpype1-ext==0.0.2
    # via localstack-core (pyproject.toml)
json5==0.10.0
    # via localstack-core (pyproject.toml)
jsondiff==2.2.1
    # via moto-ext
jsonpatch==1.33
    # via
    #   cfn-lint
    #   localstack-core
jsonpath-ng==1.7.0
    # via
    #   localstack-core (pyproject.toml)
    #   moto-ext
jsonpath-rw==1.4.0
    # via localstack-core (pyproject.toml)
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.23.0
    # via
    #   aws-sam-translator
    #   moto-ext
    #   openapi-core
    #   openapi-schema-validator
    #   openapi-spec-validator
jsonschema-path==0.3.3
    # via
    #   openapi-core
    #   openapi-spec-validator
jsonschema-specifications==2023.12.1
    # via
    #   jsonschema
    #   openapi-schema-validator
lazy-object-proxy==1.10.0
    # via openapi-spec-validator
localstack-twisted==24.3.0
    # via localstack-core
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.5.0
    # via openapi-core
moto-ext==5.0.22.post1
    # via localstack-core (pyproject.toml)
mpmath==1.3.0
    # via sympy
multipart==1.2.1
    # via moto-ext
networkx==3.4.2
    # via cfn-lint
openapi-core==0.19.4
    # via localstack-core
openapi-schema-validator==0.6.2
    # via
    #   openapi-core
    #   openapi-spec-validator
openapi-spec-validator==0.7.1
    # via
    #   moto-ext
    #   openapi-core
opensearch-py==2.8.0
    # via localstack-core (pyproject.toml)
packaging==24.2
    # via
    #   apispec
    #   build
    #   jpype1-ext
parse==1.20.2
    # via openapi-core
pathable==0.4.3
    # via jsonschema-path
plux==1.12.1
    # via
    #   localstack-core
    #   localstack-core (pyproject.toml)
ply==3.11
    # via
    #   jsonpath-ng
    #   jsonpath-rw
priority==1.3.0
    # via
    #   hypercorn
    #   localstack-twisted
psutil==6.1.1
    # via
    #   localstack-core
    #   localstack-core (pyproject.toml)
py-partiql-parser==0.5.6
    # via moto-ext
pyasn1==0.6.1
    # via rsa
pycparser==2.22
    # via cffi
pydantic==2.10.4
    # via aws-sam-translator
pydantic-core==2.27.2
    # via pydantic
pygments==2.18.0
    # via rich
pymongo==4.10.1
    # via localstack-core (pyproject.toml)
pyopenssl==24.3.0
    # via
    #   localstack-core
    #   localstack-core (pyproject.toml)
    #   localstack-twisted
pyparsing==3.2.0
    # via moto-ext
pyproject-hooks==1.2.0
    # via build
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   moto-ext
    #   opensearch-py
python-dotenv==1.0.1
    # via
    #   localstack-core
    #   localstack-core (pyproject.toml)
pyyaml==6.0.2
    # via
    #   awscli
    #   cfn-lint
    #   jsondiff
    #   jsonschema-path
    #   localstack-core
    #   localstack-core (pyproject.toml)
    #   moto-ext
    #   responses
readerwriterlock==1.0.9
    # via localstack-core
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
regex==2024.11.6
    # via cfn-lint
requests==2.32.3
    # via
    #   docker
    #   jsonschema-path
    #   localstack-core
    #   localstack-core (pyproject.toml)
    #   moto-ext
    #   opensearch-py
    #   requests-aws4auth
    #   responses
    #   rolo
requests-aws4auth==1.3.1
    # via localstack-core
responses==0.25.3
    # via moto-ext
rfc3339-validator==0.1.4
    # via openapi-schema-validator
rich==13.9.4
    # via
    #   localstack-core
    #   localstack-core (pyproject.toml)
rolo==0.7.4
    # via localstack-core
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
rsa==4.7.2
    # via awscli
s3transfer==0.10.4
    # via
    #   awscli
    #   boto3
semver==3.0.2
    # via
    #   localstack-core
    #   localstack-core (pyproject.toml)
six==1.17.0
    # via
    #   airspeed-ext
    #   jsonpath-rw
    #   python-dateutil
    #   rfc3339-validator
sympy==1.13.3
    # via cfn-lint
tailer==0.4.1
    # via
    #   localstack-core
    #   localstack-core (pyproject.toml)
typing-extensions==4.12.2
    # via
    #   aws-sam-translator
    #   cfn-lint
    #   localstack-twisted
    #   pydantic
    #   pydantic-core
    #   readerwriterlock
urllib3==2.3.0
    # via
    #   botocore
    #   docker
    #   localstack-core
    #   opensearch-py
    #   requests
    #   responses
werkzeug==3.1.3
    # via
    #   localstack-core
    #   moto-ext
    #   openapi-core
    #   rolo
wrapt==1.17.0
    # via aws-xray-sdk
wsproto==1.2.0
    # via hypercorn
xmltodict==0.14.2
    # via
    #   localstack-core
    #   moto-ext
zope-interface==7.2
    # via localstack-twisted

# The following packages are considered to be unsafe in a requirements file:
# localstack-core
# setuptools
